refactor(#204) : 컨트롤러 분리
