+++ src/main/java/odyssey/backend/presentation/auth/AuthController.java
@@ -3,13 +3,12 @@ package odyssey.backend.presentation.auth;
 import jakarta.servlet.http.HttpServletResponse;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
-import odyssey.backend.application.auth.*;
+import odyssey.backend.application.auth.LoginService;
+import odyssey.backend.application.auth.LogoutService;
+import odyssey.backend.application.auth.RefreshService;
 import odyssey.backend.domain.auth.User;
 import odyssey.backend.infrastructure.jwt.dto.response.TokenResponse;
 import odyssey.backend.presentation.auth.dto.request.LoginRequest;
-import odyssey.backend.presentation.auth.dto.request.SignUpRequest;
-import odyssey.backend.presentation.auth.dto.response.SignUpResponse;
-import odyssey.backend.presentation.auth.dto.response.UserResponse;
 import odyssey.backend.shared.response.CommonResponse;
 import odyssey.backend.shared.response.SingleCommonResponse;
 import org.springframework.http.HttpStatus;
@@ -21,21 +20,11 @@ import org.springframework.web.bind.annotation.*;
 @RequestMapping("/auth")
 public class AuthController {
 
-    private final SignUpService signUpService;
     private final LoginService loginService;
     private final LogoutService logoutService;
     private final RefreshService refreshService;
-    private final GetUserInfoService getUserInfoService;
 
     @PostMapping
-    @ResponseStatus(HttpStatus.CREATED)
-    public SingleCommonResponse<SignUpResponse> signUp(
-            @RequestBody @Valid SignUpRequest signUpRequest
-    ) {
-        return CommonResponse.ok(signUpService.signUp(signUpRequest));
-    }
-
-    @PostMapping("/login")
     @ResponseStatus(HttpStatus.OK)
     public SingleCommonResponse<TokenResponse> login(
             @RequestBody @Valid LoginRequest request,
@@ -51,7 +40,7 @@ public class AuthController {
         return CommonResponse.ok(refreshService.refreshToken(refreshToken));
     }
 
-    @DeleteMapping("/logout")
+    @DeleteMapping
     public SingleCommonResponse<String> logout(
             @AuthenticationPrincipal User user
     ) {
@@ -59,10 +48,4 @@ public class AuthController {
         return CommonResponse.ok("로그아웃되었습니다.");
     }
 
-    @GetMapping("/info")
-    public SingleCommonResponse<UserResponse> info(
-            @AuthenticationPrincipal User user
-    ) {
-        return CommonResponse.ok(getUserInfoService.getUserInfo(user));
-    }
 }
