+++ src/main/java/odyssey/backend/presentation/user/VerificationController.java
@@ -1,11 +1,11 @@
-package odyssey.backend.presentation.auth;
+package odyssey.backend.presentation.user;
 
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
-import odyssey.backend.application.auth.SendVerificationCodeService;
-import odyssey.backend.application.auth.VerificationValidUseCase;
-import odyssey.backend.presentation.auth.dto.request.SendVerificationRequest;
-import odyssey.backend.presentation.auth.dto.request.VerifyRequest;
+import odyssey.backend.application.user.SendVerificationCodeService;
+import odyssey.backend.application.user.VerificationValidUseCase;
+import odyssey.backend.presentation.user.dto.request.SendVerificationRequest;
+import odyssey.backend.presentation.user.dto.request.VerifyRequest;
 import org.springframework.web.bind.annotation.*;
 
 @RequiredArgsConstructor
