+++ src/main/java/odyssey/backend/presentation/user/dto/request/VerifyRequest.java
@@ -1,16 +1,17 @@
-package odyssey.backend.presentation.auth.dto.request;
+package odyssey.backend.presentation.user.dto.request;
 
 import jakarta.validation.constraints.NotEmpty;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
-import lombok.NoArgsConstructor;
 
 @Getter
-@NoArgsConstructor
 @AllArgsConstructor
-public class SendVerificationRequest {
+public class VerifyRequest {
 
     @NotEmpty(message = "필수값입니다.")
     private String email;
+
+    @NotEmpty(message = "필수값입니다.")
+    private String code;
 
 }
